refactor(frontend): MessageBubble を 6 ファイル に分割 (456 → 192 行)
AI チャット の メッセージ 描画 を 担う 巨大 ファイル から、 純粋 視覚 サブ コンポーネント と
Markdown ヘルパ を 抽出。 主 ファイル は 4 つ の case (image / deleted / sender / assistant) の
ルーティング に 集中。

抽出 先:
- src/components/message/MessageActionRow.tsx (73 行) — 時刻 + コピー + 削除
- src/components/message/MessageAttachmentList.tsx (51 行) — 自分の メッセージ 添付 表示
- src/components/message/MarkdownView.tsx (67 行) — react-markdown ラッパ
- src/components/message/CodeBlock.tsx (73 行) — コピー ボタン 付き コード ブロック
- src/utils/markdownContent.ts (29 行) — extractLanguage / extractTextContent

後方 互換:
- `MessageAttachmentView` 型 を MessageBubble.tsx から re-export (= MessageBubbleAi の import を 変えない)

検証:
- tsc クリーン
- eslint クリーン
- vitest 1486 件 全 PASS (185 ファイル / 回帰 なし)
- npm run build 成功
