+++ frontend/src/components/message/CodeBlock.tsx
@@ -1,0 +1,73 @@
+import { ReactNode, useMemo, useState } from 'react';
+import { ClipboardDocumentCheckIcon, ClipboardDocumentIcon } from '@heroicons/react/24/outline';
+import { extractLanguage, extractTextContent } from '../../utils/markdownContent';
+
+/**
+ * AI 応答の Markdown 内コードブロックに、言語名表示 + コピーボタン付きのヘッダを足す。
+ *
+ * ChatGPT / Claude.ai と同じ UX で、ユーザーは AI 応答のコードを 1 クリックでクリップボードに
+ * 取り込める。「メッセージ全体コピー」ボタンとは別に、コードブロック単位のコピー UI を提供する。
+ *
+ * react-markdown が rehype-highlight 後に渡してくる構造:
+ *
+ *   <pre>            ← この `pre` コンポーネント置換が CodeBlock のエントリ
+ *     <code class="language-py">
+ *       <span class="hljs-keyword">def</span>
+ *       ...
+ *     </code>
+ *   </pre>
+ *
+ * `children` の最も内側の文字列を再帰的に集めて生コードを復元する（rehype-highlight が
+ * 挿入した <span> はクラスだけ付けて textContent は変えないため、テキスト連結で正確に
+ * 元のコードが復元できる）。
+ */
+export default function CodeBlock({ children }: { children: ReactNode }) {
+  const [copied, setCopied] = useState(false);
+
+  const language = useMemo(() => extractLanguage(children), [children]);
+  const rawCode = useMemo(
+    () => extractTextContent(children).replace(/\n$/, ''),
+    [children]
+  );
+
+  const handleCopy = async () => {
+    if (!rawCode) return;
+    try {
+      await navigator.clipboard.writeText(rawCode);
+      setCopied(true);
+      window.setTimeout(() => setCopied(false), 1500);
+    } catch {
+      // clipboard API はブラウザ設定 / HTTP 環境で拒否されることがある。
+      // エラーは握りつぶし、ユーザーには反応なしのままにする（壊れた挙動より無反応の方が安全）。
+    }
+  };
+
+  return (
+    <div className="my-2 rounded-md overflow-hidden bg-[var(--color-surface-3)]">
+      <div className="flex items-center justify-between px-3 py-1 bg-[var(--color-surface-2)] border-b border-[var(--color-surface-3)] text-[11px]">
+        <span className="text-[var(--color-text-muted)] font-mono">
+          {language || 'code'}
+        </span>
+        <button
+          type="button"
+          onClick={handleCopy}
+          aria-label={copied ? 'コードをコピーしました' : 'コードをコピー'}
+          className="flex items-center gap-1 px-1.5 py-0.5 rounded text-[var(--color-text-muted)] hover:text-[var(--color-text-primary)] hover:bg-[var(--color-surface-3)] transition-colors"
+        >
+          {copied ? (
+            <>
+              <ClipboardDocumentCheckIcon className="w-3.5 h-3.5 text-green-400" />
+              <span>コピー済み</span>
+            </>
+          ) : (
+            <>
+              <ClipboardDocumentIcon className="w-3.5 h-3.5" />
+              <span>コピー</span>
+            </>
+          )}
+        </button>
+      </div>
+      <pre className="p-3 overflow-x-auto text-xs">{children}</pre>
+    </div>
+  );
+}
